refactor: Extract reusable code from `simple_https_client` to new `base` workspace member.
Note: In the next step that code will be reused by `simple_https_server`.
