+++ Cargo.toml
@@ -25,16 +25,17 @@ default = ["hacspec"]
 hacspec = ["hacspec_cryptolib"]
 evercrypt = ["evercrypt_cryptolib"]
 
-[dev-dependencies]  
+[dev-dependencies]
 dhat = "0.3.0"
 serde_json = "1.0"
-serde = {version = "1.0", features = ["derive"]}
+serde = { version = "1.0", features = ["derive"] }
 rayon = "1.3.0"
 hacspec-dev = { git = "https://github.com/hacspec/hacspec", package = "hacspec-dev" }
 criterion = "0.4"
 
 [workspace]
 members = [
+    "base",
     "simple_https_client",
     "simple_https_server",
     "bogo_shim",
