+++ base/src/stream.rs
